[SC-8] Completed task 'Slotting in new behaviours' with tests
