+++ Assets/Placeholder.txt.meta
@@ -1,0 +1,7 @@
+fileFormatVersion: 2
+guid: 90b7e26c1a1561441b0372169dac9067
+TextScriptImporter:
+  externalObjects: {}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
