+++ Packages/com.michaelrryan.topdowncharactercontroller/Runtime/Scripts/ICharacterBehaviour.cs
@@ -1,0 +1,10 @@
+﻿using System.Collections;
+using System.Collections.Generic;
+using UnityEngine;
+
+public interface ICharacterBehaviour
+{
+    void SetBehaviourUser(ICharacterController characterController);
+    ICharacterController GetBehaviourUser();
+    void Update();
+}
