+++ Packages/com.michaelrryan.topdowncharactercontroller/Runtime/Scripts/ICharacterController.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: c8be05995577aec43ae4334647c8017a
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
