+++ Packages/com.michaelrryan.topdowncharactercontroller/Tests/CustomisableBehavioursTests.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: 892b7098702ec0544aa549ea352d7e45
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
