+++ Packages/manifest.json
@@ -8,7 +8,7 @@
     "com.unity.2d.tilemap": "1.0.0",
     "com.unity.collab-proxy": "1.10.2",
     "com.unity.ide.rider": "1.2.1",
-    "com.unity.ide.visualstudio": "2.0.11",
+    "com.unity.ide.visualstudio": "2.0.12",
     "com.unity.ide.vscode": "1.2.4",
     "com.unity.test-framework": "1.1.29",
     "com.unity.textmeshpro": "2.1.6",
